"working" game player
It plays games, but it loses every one against rollout_bot which is bad.
I think the problem is in think, I'm not quite sure what to do there.
I also think that some of the functions are returning more than they should. When state and board are passed to a function, are they copies or direct references?