Working mcts_modified
tweaked mcts_vanilla a bit

mcts_modified uses rollout_bot as a basis for making rollout decisions.
Takes ages to make a single turn though

Things to modify:
In rollout_helper
- rollouts
- max_depth
Global Constants
- num_nodes
- explore_faction

Finda a balance of speed and better at playingness.
We can make vanilla as bad as we want as long as it beats random and rollout bots most of the time right?